feat: New utils (getDefaultValidityState, validate)
