+++ packages/vue/src/validate/utils/index.ts
@@ -1,3 +1,5 @@
 export * from "./create-field";
 export * from "./create-form";
 export * from "./create-rule";
+export * from "./get-default-validity-state";
+export * from "./validate";
